WIP: support for versions that do not end with -SNAPSHOT
We use feature branches with dedicated versions, having a feature-ID inside
the version name, e.g. 10.0.2-TND-9260-SNAPSHOT

Everything behind the first hyphen (-) in the Maven version is considered as a qualifier.
If that qualifier ends with -SNAPSHOT, it is treated as a snapshot version. The corresponding
OSGi version would be 10.0.2.qualifier.